Move the result comparator into its own module
